[refactor][multiarch] enhance multi-architecture support and kernel configuration
- Add x86_64 and RISC-V boot assembly implementations for complete architecture coverage
- Enhance moss_std.hpp with comprehensive freestanding standard library functionality
- Update CMake build system with improved architecture detection and configuration
- Refactor container and IPC modules for better cross-platform compatibility
- Expand kernel configuration system with architecture-specific compile-time options

+++ src/containers/slab_allocator.hpp
@@ -181,7 +181,7 @@ public:
     [[nodiscard]] double utilization() const noexcept {
         usize total = total_objects();
         if (total == 0) return 0.0;
-        return static_cast<double>(allocated_objects()) / total;
+        return static_cast<double>(allocated_objects()) / static_cast<double>(total);
     }
 
 private:
